bashWebsiteStarterTemplateGenerator
Creates files and directories for a website starter template. It is considered bad practice to generate file content in a bash script. It is advised to separate contents into their own files. I am usually an avid follower of good advice but in this case I prefer the advantage of having a single file; it is not a lot of content anyway.